// src/explorerApi/api.ts
export type RegisterType = 'R4' | 'R5' | 'R6' | 'R7' | 'R8' | 'R9';

export interface RegisterDetail {
  serializedValue: string;
  sigmaType: string;
  renderedValue: string;
}

export type AdditionalRegisters = Partial<Record<RegisterType, RegisterDetail>>;
export interface AssetInstanceInfo {
  /**
   * Token ID
   * @type {string}
   * @memberof AssetInstanceInfo
   */
  'tokenId': string;
  /**
   * Index of the asset in an output
   * @type {number}
   * @memberof AssetInstanceInfo
   */
  'index': number;
  /**
   * Amount of tokens
   * @type {number}
   * @memberof AssetInstanceInfo
   */
  'amount': bigint;
  /**
   * Name of a token
   * @type {string}
   * @memberof AssetInstanceInfo
   */
  'name'?: string;
  /**
   * Number of decimal places
   * @type {number}
   * @memberof AssetInstanceInfo
   */
  'decimals'?: number;
  /**
   * Type of a token (token standard)
   * @type {string}
   * @memberof AssetInstanceInfo
   */
  'type'?: string;
}
// ...
export interface MOutputInfo {
  /**
   * Id of the box
   * @type {string}
   * @memberof MOutputInfo
   */
  'boxId': string;
  /**
   * Id of the transaction that created the box
   * @type {string}
   * @memberof MOutputInfo
   */
  'transactionId': string;
  /**
   * Id of the block a box included in
   * @type {string}
   * @memberof MOutputInfo
   */
  'blockId'?: string;
  /**
   * Value of the box in nanoERG
   * @type {number}
   * @memberof MOutputInfo
   */
  'value': number;
  /**
   * Index of the output in a transaction
   * @type {number}
   * @memberof MOutputInfo
   */
  'index': number;
  /**
   * Global index of the output in the blockchain
   * @type {number}
   * @memberof MOutputInfo
   */
  'globalIndex'?: number;
  /**
   * Height at which the box was created
   * @type {number}
   * @memberof MOutputInfo
   */
  'creationHeight': number;
  /**
   * Height at which the box got fixed in blockchain
   * @type {number}
   * @memberof MOutputInfo
   */
  'settlementHeight'?: number;
  /**
   * Serialized ergo tree
   * @type {string}
   * @memberof MOutputInfo
   */
  'ergoTree': string;
  /**
   * An address derived from ergo tree
   * @type {string}
   * @memberof MOutputInfo
   */
  'address': string;
  /**
   *
   * @type {Array<AssetInstanceInfo>}
   * @memberof MOutputInfo
   */
  'assets'?: Array<AssetInstanceInfo>;
  /**
   *
   * @type {{ [key: string]: string; }}
   * @memberof MOutputInfo
   */
  'additionalRegisters': AdditionalRegisters;
  /**
   * Id of the transaction this output was spent by
   * @type {string}
   * @memberof MOutputInfo
   */
  'spentTransactionId'?: string;
  /**
   *
   * @type {boolean}
   * @memberof MOutputInfo
   */
  'mainChain': boolean;
}
// ...
export interface OutputInfo {
  /**
   * Id of the box
   * @type {string}
   * @memberof OutputInfo
   */
  'boxId': string;
  /**
   * Id of the transaction that created the box
   * @type {string}
   * @memberof OutputInfo
   */
  'transactionId': string;
  /**
   * Id of the block a box included in
   * @type {string}
   * @memberof OutputInfo
   */
  'blockId': string;
  /**
   * Value of the box in nanoERG
   * @type {number}
   * @memberof OutputInfo
   */
  'value': number;
  /**
   * Index of the output in a transaction
   * @type {number}
   * @memberof OutputInfo
   */
  'index': number;
  /**
   * Global index of the output in the blockchain
   * @type {number}
   * @memberof OutputInfo
   */
  'globalIndex': number;
  /**
   * Height at which the box was created
   * @type {number}
   * @memberof OutputInfo
   */
  'creationHeight': number;
  /**
   * Height at which the box got fixed in blockchain
   * @type {number}
   * @memberof OutputInfo
   */
  'settlementHeight': number;
  /**
   * Serialized ergo tree
   * @type {string}
   * @memberof OutputInfo
   */
  'ergoTree': string;
  /**
   *
   * @type {string}
   * @memberof OutputInfo
   */
  'ergoTreeConstants': string;
  /**
   *
   * @type {string}
   * @memberof OutputInfo
   */
  'ergoTreeScript': string;
  /**
   * An address derived from ergo tree
   * @type {string}
   * @memberof OutputInfo
   */
  'address': string;
  /**
   *
   * @type {Array<AssetInstanceInfo>}
   * @memberof OutputInfo
   */
  'assets'?: Array<AssetInstanceInfo>;
  /**
   *
   * @type {{ [key: string]: string; }}
   * @memberof OutputInfo
   */
  'additionalRegisters': AdditionalRegisters;
  /**
   * Id of the transaction this output was spent by
   * @type {string}
   * @memberof OutputInfo
   */
  'spentTransactionId'?: string;
  /**
   *
   * @type {boolean}
   * @memberof OutputInfo
   */
  'mainChain': boolean;
}